feat(core): basic T1 homescreen
